requirements.txt
fastapi
uvicorn[standard]
requests
azure-cognitiveservices-speech
pydantic
